+++ wemake_python_styleguide/violations/consistency.py
@@ -322,7 +322,6 @@ class MultipleIfsInComprehensionViolation(ASTViolation):
     code = 307
 
 
-# TODO: move to `best_practices.py`
 class ConstantComparisonViolation(ASTViolation):
     """
     Forbids to have comparisons between two literals.
@@ -357,6 +356,38 @@ class ConstantComparisonViolation(ASTViolation):
     code = 308
 
 
+class ComparisonOrderViolation(ASTViolation):
+    """
+    Forbids comparision where argument doesn't come first.
+
+    Reasoning:
+        It is hard to read the code when
+        you have to shuffle ordering of the arguments all the time.
+        Bring a consistency to the comparison!
+
+    Solution:
+        Refactor your comparison expression, place the argument first.
+
+    Example::
+
+        # Correct:
+        if some_x > 3:
+        if 3 < some_x < 10:
+
+        # Wrong:
+        if 3 < some_x:
+
+    Note:
+        Returns Z309 as error code
+
+    """
+
+    should_use_text = False
+    #: Error message shown to the user.
+    error_template = 'Found reversed comparison order'
+    code = 309
+
+
 class BadNumberSuffixViolation(TokenizeViolation):
     """
     Forbids to use capital ``X``, ``O``, ``B``, and ``E`` in numbers.
@@ -387,42 +418,10 @@ class BadNumberSuffixViolation(TokenizeViolation):
         number_with_scientific_notation = 1.5E+10
 
     Note:
-        Returns Z309 as error code
-
-    """
-
-    code = 309
-    #: Error message shown to the user.
-    error_template = 'Found underscored number: {0}'
-
-
-class ComparisonOrderViolation(ASTViolation):
-    """
-    Forbids comparision where argument doesn't come first.
-
-    Reasoning:
-        It is hard to read the code when
-        you have to shuffle ordering of the arguments all the time.
-        Bring a consistency to the comparison!
-
-    Solution:
-        Refactor your comparison expression, place the argument first.
-
-    Example::
-
-        # Correct:
-        if some_x > 3:
-        if 3 < some_x < 10:
-
-        # Wrong:
-        if 3 < some_x:
-
-    Note:
         Returns Z310 as error code
 
     """
 
-    should_use_text = False
     #: Error message shown to the user.
-    error_template = 'Found reversed comparison order'
+    error_template = 'Found underscored number: {0}'
     code = 310