+++ wemake_python_styleguide/visitors/presets/general.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
 
-from wemake_python_styleguide.visitors.ast import keywords, naming, numbers
-from wemake_python_styleguide.visitors.ast.classes import WrongClassVisitor
-from wemake_python_styleguide.visitors.ast.comparisons import (
-    ConstantComparisonVisitor,
+from wemake_python_styleguide.visitors.ast import (
+    comparisons,
+    keywords,
+    naming,
+    numbers,
 )
+from wemake_python_styleguide.visitors.ast.classes import WrongClassVisitor
 from wemake_python_styleguide.visitors.ast.functions import (
     WrongFunctionCallVisitor,
 )
@@ -12,7 +14,6 @@ from wemake_python_styleguide.visitors.ast.imports import WrongImportVisitor
 from wemake_python_styleguide.visitors.ast.modules import (
     EmptyModuleContentsVisitor,
 )
-from wemake_python_styleguide.visitors.ast.order import WrongOrderVisitor
 from wemake_python_styleguide.visitors.ast.strings import WrongStringVisitor
 from wemake_python_styleguide.visitors.filenames.wrong_module_name import (
     WrongModuleNameVisitor,
@@ -23,19 +24,24 @@ GENERAL_PRESET = (
     # General:
     keywords.WrongRaiseVisitor,
     keywords.WrongKeywordVisitor,
+    keywords.WrongListComprehensionVisitor,
+
     WrongFunctionCallVisitor,
     WrongImportVisitor,
+
     naming.WrongNameVisitor,
     naming.WrongModuleMetadataVisitor,
+
     numbers.MagicNumberVisitor,
     WrongStringVisitor,
-    EmptyModuleContentsVisitor,
-    ConstantComparisonVisitor,
-    WrongOrderVisitor,
+
+    comparisons.ConstantComparisonVisitor,
+    comparisons.WrongOrderVisitor,
 
     # Classes:
     WrongClassVisitor,
 
     # Modules:
     WrongModuleNameVisitor,
+    EmptyModuleContentsVisitor,
 )